+++ R/LearnerClassifDebug.R
@@ -105,6 +105,27 @@ LearnerClassifDebug = R6Class("LearnerClassifDebug", inherit = LearnerClassif,
     #'   Additional arguments passed to [`unmarshal_model()`].
     unmarshal = function(...) {
       learner_unmarshal(.learner = self, ...)
+    },
+
+    #' @description
+    #' Returns 0 for each feature seen in training.
+    #' @return Named `numeric()`.
+    importance = function() {
+      if (is.null(self$model)) {
+        stopf("No model stored")
+      }
+      fns = self$state$feature_names
+      set_names(rep(0, length(fns)), fns)
+    },
+
+    #' @description
+    #' Always returns character(0).
+    #' @return `character()`.
+    selected_features = function() {
+      if (is.null(self$model)) {
+        stopf("No model stored")
+      }
+      character(0)
     }
   ),
   active = list(
@@ -169,8 +190,15 @@ LearnerClassifDebug = R6Class("LearnerClassifDebug", inherit = LearnerClassif,
         stopf("Early stopping is only possible when a validation task is present.")
       }
 
-      model = list(response = as.character(sample(task$truth(), 1L)), pid = Sys.getpid(), id = UUIDgenerate(),
-        random_number = sample(100000, 1), iter = if (isTRUE(pv$early_stopping)) sample(pv$iter %??% 1L, 1L) else pv$iter %??% 1L
+      model = list(
+          response = as.character(sample(task$truth(), 1L)),
+          pid = Sys.getpid(),
+          id = UUIDgenerate(),
+          random_number = sample(100000, 1),
+          iter = if (isTRUE(pv$early_stopping))
+            sample(pv$iter %??% 1L, 1L)
+          else
+            pv$iter %??% 1L
       )
 
       if (!is.null(valid_truth)) {